convert to f64

+++ source/rust/vndf/src/shared/physics/collision.rs
@@ -10,8 +10,8 @@ use client::graphics::SHIP_SIZE;
 // then we'd need this back, at a minimum
 #[derive(Clone, Debug, RustcDecodable, RustcEncodable)]
 pub enum CollideKind {
-    Ship(Convex<Pnt2<f32>>),
-    Planet(Ball<f32>),
+    Ship(Convex<Pnt2<f64>>),
+    Planet(Ball<f64>),
 }
 
 #[derive(Clone, Debug, RustcDecodable, RustcEncodable)]
@@ -27,7 +27,7 @@ impl Collider {
     /// builds based on current ship mesh layout (from equilateral triangle)
     // TODO: make ship mesh points as public in shapes module
     pub fn new_from_ship (scaling_factor: f32) -> Collider {
-	let size = SHIP_SIZE/2.0 * scaling_factor;
+	let size = (SHIP_SIZE/2.0 * scaling_factor) as f64;
 	let p = vec![Pnt2::new(-0.5, -0.5) * size,
 		     Pnt2::new( 0.5, -0.5) * size,
 		     Pnt2::new( 0.0,  0.5) * size,];
@@ -37,7 +37,7 @@ impl Collider {
     }
 
     pub fn new_from_planet (planet_size: f32, scaling_factor: f32) -> Collider {
-	let size = planet_size/2.0 * scaling_factor;
+	let size = (planet_size/2.0 * scaling_factor) as f64;
 	let b = Ball::new(size);
 	
 	Collider::new(CollideKind::Planet(b))
@@ -45,8 +45,8 @@ impl Collider {
 
     /// requires two colliders and their associated positions in the world
     pub fn check_collision (&self,
-                            pos: &Vec2<f32>,
-                            other: (&Collider,&Vec2<f32>))
+                            pos: &Vec2<f64>,
+                            other: (&Collider,&Vec2<f64>))
                             -> bool {
         let (other_kind, other_pos) = (&other.0.kind,other.1);
 	let mut is_collide = false;
